🚚 move tests to a __tests__ folder next to their source instad of a separate tests folder at the root.
More inline with what we usually do, and i was having config issues when it was in a separate folder.
