+++ src/utils/__tests__/createRequestHeaders.test.ts
@@ -1,4 +1,4 @@
-import createRequestHeaders from '../../src/utils/createRequestHeaders';
+import createRequestHeaders from '../createRequestHeaders';
 
 describe(`create header object`, () => {
   it(`should be able to pass in a sync getAccessToken function`, async () => {
